+++ backend/src/main/java/com/etirovaf/backend/member/presentation/MemberController.java
@@ -1,11 +1,8 @@
 package com.etirovaf.backend.member.presentation;
 
-import com.etirovaf.backend.auth.application.AuthService;
-import com.etirovaf.backend.auth.model.dto.request.SignupRequest;
+import com.etirovaf.backend.member.model.dto.request.SignupRequest;
 import com.etirovaf.backend.common.domain.ResponseHandler;
 import com.etirovaf.backend.member.application.MemberService;
-import com.etirovaf.backend.member.model.dto.request.MemberInfo;
-import com.etirovaf.backend.member.model.entity.Member;
 import com.etirovaf.backend.member.model.entity.Role;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -21,18 +18,17 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "02.회원")
 public class MemberController {
     private final MemberService service;
-    private final AuthService authService;
 
     @PostMapping("/add")
     @Operation(summary="회원가입", description = "회원을 목록에 저장한다.", tags={"02.회원",})
-    public ResponseEntity<ResponseHandler<Boolean>> addMember(@RequestBody Member member){
+    public ResponseEntity<ResponseHandler<Boolean>> addMember(@RequestBody SignupRequest signupRequest){
         log.info("addMember");
-        member.setRole(Role.MEMBER);
+        signupRequest.setRole(Role.MEMBER);
         return ResponseEntity
                 .ok()
                 .body(ResponseHandler.<Boolean>builder()
                         .message("회원가입 성공")
-                        .data(authService.addMember(SignupRequest.of(member)))
+                        .data(service.addMember(signupRequest))
                         .build()
                 );
     }
